Fix Landing Page Build
upgrading next caused some breaking changes with static site generation
