+++ apps/landing/project.json
@@ -3,56 +3,13 @@
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "apps/landing",
   "projectType": "application",
-  "generators": {},
   "targets": {
-    "build": {
-      "executor": "@nx/next:build",
-      "options": {
-        "outputPath": "dist/apps/landing"
-      },
-      "configurations": {
-        "production": {},
-        "development": {
-          "outputPath": "apps/landing"
-        }
-      },
-      "outputs": ["{options.outputPath}"],
-      "defaultConfiguration": "production"
-    },
-    "serve": {
-      "executor": "@nx/next:server",
-      "options": {
-        "buildTarget": "landing:build",
-        "port": 4300,
-        "dev": true
-      },
-      "configurations": {
-        "production": {
-          "buildTarget": "landing:build:production",
-          "dev": false
-        },
-        "development": {
-          "buildTarget": "landing:build:development",
-          "dev": true
-        }
-      },
-      "defaultConfiguration": "development"
-    },
-    "export": {
-      "executor": "@nx/next:export",
-      "options": {
-        "buildTarget": "landing:build:production"
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    },
     "test": {
       "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
         "jestConfig": "apps/landing/jest.config.ts"
-      },
-      "outputs": ["{workspaceRoot}/coverage/apps/landing"]
+      }
     }
   },
   "tags": []
