+++ package.json
@@ -13,12 +13,12 @@
     "yarn": ">=1.22.21 <2"
   },
   "scripts": {
-    "init:project": "ts-node ./scripts/init-project.ts",
+    "init:project": "zx ./scripts/init-project.mjs",
     "nx": "nx",
     "start": "nx serve",
     "start:api": "nx serve api",
     "start:worker": "nx serve jetstream-worker",
-    "start:landing": "nx serve landing",
+    "start:landing": "nx dev landing",
     "start:docs": "cd apps/docs && yarn start",
     "start:electron": "nx serve electron-jetstream",
     "start:electron-worker": "nx build electron-worker --configuration=development --watch",
@@ -54,7 +54,7 @@
     "build:cron": "yarn db:generate && nx build cron-tasks --prod && yarn generate:version",
     "build:api:test": "nx build api --configuration=test",
     "build:sw": "nx build download-zip-sw --prod",
-    "build:landing": "cross-env NODE_ENV=production nx run landing:export --skip-nx-cache && yarn build:landing:sitemap",
+    "build:landing": "cross-env NODE_ENV=production nx run landing:build --skip-nx-cache && yarn build:landing:sitemap",
     "build:landing:sitemap": "next-sitemap",
     "build:docs": "cd apps/docs && yarn build",
     "build:storybook": "storybook build -c libs/ui/.storybook -o dist/storybook -s node_modules/@salesforce-ux/design-system/assets/styles",
@@ -64,13 +64,13 @@
     "electron:extract:asar-x86": "npx asar extract dist/executables/mac/Jetstream.app/Contents/Resources/app.asar dist/executables/mac/extracted",
     "electron:extract:asar-arm": "npx asar extract dist/executables/mac-arm64/Jetstream.app/Contents/Resources/app.asar dist/executables/mac-arm64/extracted",
     "release": "dotenv release-it ${0}",
-    "release:build": "ts-node ./scripts/build-release.ts",
+    "release:build": "zx ./scripts/build-release.mjs",
     "release:electron": "yarn build:electron && concurrently --group --restart-tries=3 \"npm:electron:dist:*\"",
     "bundle-analyzer:client": "npx webpack-bundle-analyzer dist/apps/jetstream/stats.json",
     "docs:index": "docker run -it -e APPLICATION_ID=21D7I5RB7N -e API_KEY=$ALGOLIA_API_KEY -e \"CONFIG=$(cat ./apps/docs/algolia-config.json | jq -r tostring)\" algolia/docsearch-scraper",
-    "rollbar:upload-sourcemaps": "yarn generate:version && ts-node ./scripts/upload-source-maps.ts",
+    "rollbar:upload-sourcemaps": "yarn generate:version && nx ./scripts/upload-source-maps.mjs",
     "generate:data-table-css": "nx workspace-generator convert-sass-to-css ui --filename /src/lib/data-table/data-table-styles.scss",
-    "generate:version": "ts-node ./scripts/generate-version.ts",
+    "generate:version": "zx ./scripts/generate-version.mjs",
     "db:generate": "prisma generate",
     "db:migrate": "prisma migrate deploy",
     "db:seed": "chmod +x prisma/seed.sh && prisma db seed",
@@ -139,7 +139,7 @@
     "@tailwindcss/forms": "^0.5.2",
     "@tailwindcss/typography": "^0.5.4",
     "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "13.4.0",
+    "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "^13.0.7",
     "@types/amplitude-js": "^7.0.0",
     "@types/async": "^3.2.3",
@@ -194,11 +194,11 @@
     "electron-updater": "^5.0.1",
     "esbuild": "^0.19.2",
     "eslint": "8.48.0",
-    "eslint-config-next": "^13.2.4",
+    "eslint-config-next": "14.0.4",
     "eslint-config-prettier": "9.0.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-react": "7.31.11",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-storybook": "^0.6.12",
     "exitzero": "^1.0.1",
@@ -208,7 +208,7 @@
     "jest-environment-node": "^29.4.1",
     "jsdom": "22.1.0",
     "jsonc-eslint-parser": "^2.1.0",
-    "next-sitemap": "^2.0.11",
+    "next-sitemap": "^4.2.3",
     "node-fetch": "^2.6.7",
     "npm-run-all": "^4.1.5",
     "nx": "18.0.7",
@@ -219,7 +219,7 @@
     "prisma": "^5.9.1",
     "react-refresh": "^0.10.0",
     "release-it": "^15.6.0",
-    "sass": "1.55.0",
+    "sass": "1.62.1",
     "sass-loader": "^10.0.4",
     "start-server-and-test": "^1.15.2",
     "storybook": "7.0.20",
@@ -238,7 +238,7 @@
     "webpack": "5.88.1",
     "webpack-merge": "^5.8.0",
     "worker-loader": "^2.0.0",
-    "zx": "^4.3.0"
+    "zx": "^7.2.3"
   },
   "dependencies": {
     "@emotion/react": "11.11.1",
